add functions for different uses
add functions for each to fetch data from the DOM, create object for the todo , add object to the array, add object to the library and to edit the todo when user chose to edit. these functions should work properly on their own. set up the required skeleton files for the next step.